requirements.txt
awscli==1.35.12
awsebcli==3.21.0
awslambdaric==2.0.10
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
boto3==1.35.2
botocore==1.35.46
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
docutils==0.16
exceptiongroup==1.2.2
execnet==2.1.1
filelock==3.16.1
Flask==2.0.3
fsspec==2024.10.0
hel==0.1.0
huggingface-hub==0.26.2
idna==3.7
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pathspec==0.10.1
pika==1.3.2
pillow==11.0.0
pip-system-certs==4.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pymongo==4.10.1
pytest==8.3.3
pytest-mock==3.14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rsa==4.7.2
s3transfer==0.10.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.2.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.6
sqldf==0.4.2
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tldextract==5.1.2
tokenizers==0.20.1
tomli==2.0.2
torch==2.5.1
tornado==6.4
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.16.0
zipp==3.20.2
